move flag parsing to cli app package
